Code recleanup
I think Aditi bulldozed over my old commits when merging her branch with remote, so this merge readds those commits and makes it compatible with the new changes.

+++ src/main/java/frc/robot/ControllableSubsystem.java
@@ -4,6 +4,6 @@ public interface ControllableSubsystem {
     double getTotalCurrentDrawn();
     double minCurrent();
 
-    void setCurrentLimit(int limit);
+    void setCurrentLimit(double limit);
     int getCurrentLimit();
 }
